fixed syscall error handling to pass syserr tests 1,2,3,5,8. test 6 still failing but waiting on development of vidmap

+++ student-distrib/syscall_funcs.h
@@ -11,6 +11,6 @@ void parse_input(const char * in, char * exec_buf, char * args_buf, int size);
 int syscall_open(const char * name);
 int syscall_close(int fd);
 int syscall_getargs(char * buf, int nbytes);
-int syscall_vidmap(int ** vid_ptr);
+int syscall_vidmap(uint8_t ** vid_ptr);
 #endif /*ASM*/
 #endif /*_SYSCALL_FUNCS_H*/
